refactor(extensions): move TracingExporter to src/extensions/tracing/, rename ext-opentelemetry → ext-tracing-opentelemetry
- src/extensions/tracing/{tracing-exporter,index}.ts
- src/extensions/interfaces/index.ts shim re-exports from ../tracing/
- extensions/ext-tracing-opentelemetry (@veryfront/ext-tracing-opentelemetry) maps veryfront/extensions/tracing
- Root deno.json workspace + ./extensions/tracing export + imports alias
- Consumer stragglers (integration.test.ts test names + virtual:// origins, recommendations.ts, bootstrap.ts, observability/* comments, runtime name) updated
