+++ extensions/ext-tracing-opentelemetry/deno.json
@@ -1,5 +1,5 @@
 {
-  "name": "@veryfront/ext-opentelemetry",
+  "name": "@veryfront/ext-tracing-opentelemetry",
   "version": "0.1.0",
   "exports": "./src/index.ts",
   "veryfront": {
@@ -28,8 +28,8 @@
     "@opentelemetry/semantic-conventions": "npm:@opentelemetry/semantic-conventions@1.40.0",
     "@std/assert": "jsr:@std/assert@1",
     "@std/testing/bdd": "jsr:@std/testing@1/bdd",
-    "veryfront/extensions": "../../../src/extensions/index.ts",
-    "veryfront/extensions/interfaces": "../../../src/extensions/interfaces/index.ts"
+    "veryfront/extensions": "../../src/extensions/index.ts",
+    "veryfront/extensions/tracing": "../../src/extensions/tracing/index.ts"
   },
   "tasks": {
     "test": "deno test --no-check --allow-all src/"
