+++ src/extensions/tracing/index.ts
@@ -1,0 +1,11 @@
+/**
+ * Tracing category barrel — TracingExporter contract.
+ *
+ * @module extensions/tracing
+ */
+
+export type {
+  SpanData,
+  TracerProvider,
+  TracingExporter,
+} from "./tracing-exporter.ts";
