+++ src/server/bootstrap.ts
@@ -7,7 +7,7 @@ import { AIProviderRegistryName, createAIProviderRegistry } from "#veryfront/ext
 import { createBuiltinExtensions } from "#veryfront/extensions/builtin-extensions.ts";
 import { MISSING_EXTENSION_ERROR } from "#veryfront/extensions/errors.ts";
 import { getRecommendation } from "#veryfront/extensions/recommendations.ts";
-import type { TracingExporter } from "#veryfront/extensions/interfaces/tracing-exporter.ts";
+import type { TracingExporter } from "#veryfront/extensions/tracing/tracing-exporter.ts";
 import {
   setGlobalActiveSpanAccessor,
   setGlobalMetricsAPI,
